Adapted compilation database to clang requirements
In "file" field bake is now putting absolute paths of source files. In
the projects with nested source directories, "directory" shows the main
project dir, while the rest of the path is passed with the file name.
Clang doesn't support having relative subdirectories in the "file"
field, so in order to keep 'directory'and 'command' fields untouched, we
need to switch to absolute file paths.

Also the default file name for compilation database has been changed to
'compile_commands.json' - this is the default name used by clang.

Adapted unit tests, reduced code duplication.