Add generic bake args to collections per project
Advanced bake configuration cannot go without adaptions. The adapt
feature make bake quite unusable when you have to remember all the
individual adapts and maybe even type them. This is the reason why
almost all bake users write wrapper like shell scripts or even rake
configurations in order to store the long bake invocations somewhere.
This patch add a option to bakery that is part of the bake toolkit that
allows to define bake arguments for a collection. Normally collections
are used to run multiple bake call, but with this extension a collection
may act as a wrapper that is "bake aware".